delete older files 14 days

+++ 25.DeleteOlderFiles.sh
@@ -18,7 +18,10 @@ Files=$(find . -name "*.log" -type f -mtime +14)
 echo $Files
 
 for File in $Files
-do  
-  rm $File
-  echo "Delete file $File"
+do   
+  if [ -z "$Files"]; then
+  echo -e "$Y There are Files to delete older than 14 days"
+  else
+    rm $File
+    echo "Deleted file $File"
 done
